+++ src/index.ts
@@ -1,1 +1,1 @@
-export * from './messages';
+export * from './messages/formatters';
